Update TwoFactorAuth.java
Okay I've initialised the BufferedReader and added the ISR - I'm not sure if the codeInput method is correct now that it's expected to take an argument. These were the suggested changes in VS code and while I know it may not be perfect, it eliminated the errors!

As it's late here and I know we have a deadline to meet, feel free to make any further changes over the next 8 hours or so and I'll catch up with anything else that needs doing in the morning.